feat: вывел работу с localStorage в отдельные методы

+++ src/features/Logout/index.ts
@@ -1,6 +1,9 @@
 'use client'
 
 export const Logout = () => {
-    localStorage.clear()
-    document.location.reload()
+
+    if (typeof window !== 'undefined') {
+        localStorage.clear()
+        document.location.reload()
+    }
 }
